+++ lib/asset.dart
@@ -1,36 +1,40 @@
+import 'dart:async';
+
 import 'package:flutter/services.dart';
 import 'package:multi_image_picker/picker.dart';
 
 class Asset {
-  String identifier;
-  int width;
-  int height;
+  String _identifier;
+  ByteData thumbData;
   ByteData imageData;
 
-  void requestThumbnail(
-      int width, int height, void handler(ByteData imageData)) {
-    const prefix = 'multi_image_picker/image';
-    var channelName = prefix + '/' + this.identifier;
+  Asset(this._identifier);
 
-    BinaryMessages.setMessageHandler(channelName, (message) {
-      this.imageData = message;
-      handler(message);
-      BinaryMessages.setMessageHandler(channelName, null);
-    });
-
-    MultiImagePicker.requestThumbnail(this.identifier, width, height);
+  String get _channel {
+    return 'multi_image_picker/image/$_identifier';
   }
 
-  void requestOriginal(void handler(ByteData imageData)) {
-    const prefix = 'multi_image_picker/image';
-    var channelName = prefix + '/' + this.identifier;
-
-    BinaryMessages.setMessageHandler(channelName, (message) {
-      this.imageData = message;
-      handler(message);
-      BinaryMessages.setMessageHandler(channelName, null);
+  Future<dynamic> requestThumbnail(int width, int height) {
+    Completer completer = new Completer();
+    BinaryMessages.setMessageHandler(_channel, (ByteData message) {
+      thumbData = message;
+      completer.complete(message);
+      BinaryMessages.setMessageHandler(_channel, null);
     });
 
-    MultiImagePicker.requestOriginal(this.identifier);
+    MultiImagePicker.requestThumbnail(_identifier, width, height);
+    return completer.future;
+  }
+
+  Future<dynamic> requestOriginal() {
+    Completer completer = new Completer();
+    BinaryMessages.setMessageHandler(_channel, (ByteData message) {
+      imageData = message;
+      completer.complete(message);
+      BinaryMessages.setMessageHandler(_channel, null);
+    });
+
+    MultiImagePicker.requestOriginal(_identifier);
+    return completer.future;
   }
 }